+++ pkg/cgroup/cgroupv1/blockio_test.go
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package cgroup_test
+package cgroupv1_test
 
 import (
 	"fmt"
@@ -19,25 +19,31 @@ import (
 
 	"github.com/adalton/teleport-exercise/pkg/adaptation/os"
 	"github.com/adalton/teleport-exercise/pkg/adaptation/os/ostest"
-	"github.com/adalton/teleport-exercise/pkg/cgroup/v1"
+	"github.com/adalton/teleport-exercise/pkg/cgroup/cgroupv1"
 
 	"github.com/stretchr/testify/assert"
 )
 
-func Test_cpu_Apply(t *testing.T) {
+func Test_blkio_Apply(t *testing.T) {
 	path := "/sys/fs/cgroup/jobs/889f7cc2-9935-4773-aaa1-b94478abc923"
 	writeRecorder := ostest.WriteFileMock{}
 	adapter := &os.Adapter{
 		WriteFileFn: writeRecorder.WriteFile,
 	}
 
-	cpu := cgroup.NewCpuControllerDetailed(adapter).SetCpus(2.0)
-	cpu.Apply(path)
+	readBps := "1:2 1G"
+	writeBps := "1:3 900M"
+
+	blkio := cgroupv1.NewBlockIoControllerDetailed(adapter).
+		SetReadBpsDevice(readBps).
+		SetWriteBpsDevice(writeBps)
+
+	blkio.Apply(path)
 
 	assert.Equal(t, 2, len(writeRecorder.Events))
-	assert.Equal(t, fmt.Sprintf("%s/%s", path, cgroup.CpuPeriodFilename), writeRecorder.Events[0].Name)
-	assert.Equal(t, []byte("100000"), writeRecorder.Events[0].Data)
+	assert.Equal(t, fmt.Sprintf("%s/%s", path, cgroupv1.BlkioThrottleReadBpsDevice), writeRecorder.Events[0].Name)
+	assert.Equal(t, []byte(readBps), writeRecorder.Events[0].Data)
 
-	assert.Equal(t, fmt.Sprintf("%s/%s", path, cgroup.CpuQuotaFilename), writeRecorder.Events[1].Name)
-	assert.Equal(t, []byte("200000"), writeRecorder.Events[1].Data)
+	assert.Equal(t, fmt.Sprintf("%s/%s", path, cgroupv1.BlkioThrottleWriteBpsDevice), writeRecorder.Events[1].Name)
+	assert.Equal(t, []byte(writeBps), writeRecorder.Events[1].Data)
 }
