+++ pkg/cgroup/cgroupv1/cgroupv1test/controller_mock.go
@@ -11,6 +11,20 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package cgrouptest provides a collection of components to help with
-// unit testing clients of the cgroup/v1 package.
-package cgrouptest
+package cgroupv1test
+
+// ControllerMock provides a mock implementation of the v1.Controller interface
+// for use in unit test.  This implementation does not modify any actual
+// cgroup.
+type ControllerMock struct {
+	ControllerName   string
+	ApplyReturnValue error
+}
+
+func (d *ControllerMock) Name() string {
+	return d.ControllerName
+}
+
+func (d *ControllerMock) Apply(string) error {
+	return d.ApplyReturnValue
+}
