+++ pkg/cgroup/cgroupv1/cgroupv1test/doc.go
@@ -11,15 +11,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package cgroup
-
-// Controller defines the interface to a cgroup controller -- objects that
-// model concrete cgroup controlers and their configuration options.
-type Controller interface {
-	// Name returns the name of the cgroup
-	Name() string
-
-	// Apply applies this controller's configuration to the cgroup at the
-	// given path.
-	Apply(path string) error
-}
+// Package cgroupv1test provides a collection of components to help with
+// unit testing clients of the cgroupv1 package.
+package cgroupv1test
