+++ pkg/cgroup/cgroupv1/doc.go
@@ -11,20 +11,12 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package cgrouptest
-
-// ControllerMock provides a mock implementation of the v1.Controller interface
-// for use in unit test.  This implementation does not modify any actual
-// cgroup.
-type ControllerMock struct {
-	ControllerName   string
-	ApplyReturnValue error
-}
-
-func (d *ControllerMock) Name() string {
-	return d.ControllerName
-}
-
-func (d *ControllerMock) Apply(string) error {
-	return d.ApplyReturnValue
-}
+// Package cgroupv1 provides an simple abstraction over the cgroup v1 interface.
+// The rationale to implement was driven by the fact that on my system, both
+// the v1 and v2 interfaces are mounted and the v1 interface is in use.
+//
+// I've versioned the package structure so that we could easily build a v2
+// implementation.  We could also add some code to the parent package to
+// enable us to examine the system and automatically pick the most suitable
+// implementation.
+package cgroupv1
