+++ test/job/cpulimit/cpulimit.go
@@ -17,11 +17,11 @@ import (
 	"fmt"
 	"runtime"
 
-	"github.com/adalton/teleport-exercise/pkg/cgroup/v1"
+	"github.com/adalton/teleport-exercise/pkg/cgroup/cgroupv1"
 	"github.com/adalton/teleport-exercise/pkg/jobmanager"
 )
 
-func runTest(controllers ...cgroup.Controller) {
+func runTest(controllers ...cgroupv1.Controller) {
 
 	job := jobmanager.NewJob("theOwner", "my-test", controllers,
 		"/usr/bin/stress-ng",
@@ -72,5 +72,5 @@ func main() {
 	runTest()
 
 	fmt.Println("Running CPU test with cgroup constraints at 0.5 CPU")
-	runTest(cgroup.NewCpuController().SetCpus(0.5))
+	runTest(cgroupv1.NewCpuController().SetCpus(0.5))
 }
